+++ webui/react/src/components/CheckpointModalTrigger.test.tsx
@@ -18,15 +18,11 @@ vi.mock('services/api', () => ({
   },
 }));
 
-jest.mock('index', () => {
-  return {
-    router: {
-      navigate: (path: string) => {
-        global.window.history.pushState({}, '', path);
-      },
-    },
-  };
-});
+jest.mock('router', () => ({
+  navigate: (path: string) => {
+    global.window.history.pushState({}, '', path);
+  },
+}));
 
 const user = userEvent.setup();
