+++ webui/react/src/components/ResourcePoolCard.test.tsx
@@ -15,15 +15,11 @@ vi.mock('services/api', () => ({
   getResourcePools: () => Promise.resolve({}),
 }));
 
-jest.mock('index', () => {
-  return {
-    router: {
-      navigate: (path: string) => {
-        global.window.history.pushState({}, '', path);
-      },
-    },
-  };
-});
+jest.mock('router', () => ({
+  navigate: (path: string) => {
+    global.window.history.pushState({}, '', path);
+  },
+}));
 
 const setup = (pool: ResourcePool) => {
   const view = render(
