+++ webui/react/src/components/kit/LogViewer/LogViewer.test.tsx
@@ -17,15 +17,11 @@ interface TestLog {
   time: string;
 }
 
-jest.mock('index', () => {
-  return {
-    router: {
-      navigate: (path: string) => {
-        global.window.history.pushState({}, '', path);
-      },
-    },
-  };
-});
+jest.mock('router', () => ({
+  navigate: (path: string) => {
+    global.window.history.pushState({}, '', path);
+  },
+}));
 
 const DEFAULT_MIN_WORD_COUNT = 5;
 const DEFAULT_MAX_WORD_COUNT = 8;
