+++ webui/react/src/hooks/useModal/Checkpoint/useModalCheckpoint.test.tsx
@@ -19,15 +19,11 @@ vi.mock('services/api', () => ({
   },
 }));
 
-jest.mock('index', () => {
-  return {
-    router: {
-      navigate: (path: string) => {
-        global.window.history.pushState({}, '', path);
-      },
-    },
-  };
-});
+jest.mock('router', () => ({
+  navigate: (path: string) => {
+    global.window.history.pushState({}, '', path);
+  },
+}));
 
 const { experiment, checkpoint } = generateTestExperimentData();
