+++ webui/react/src/hooks/useModal/UserSettings/useModalCreateUser.test.tsx
@@ -21,15 +21,11 @@ vi.mock('services/api', () => ({
   postUser: vi.fn().mockReturnValue({ user: { id: 1 } }),
 }));
 
-jest.mock('index', () => {
-  return {
-    router: {
-      navigate: (path: string) => {
-        global.window.history.pushState({}, '', path);
-      },
-    },
-  };
-});
+jest.mock('router', () => ({
+  navigate: (path: string) => {
+    global.window.history.pushState({}, '', path);
+  },
+}));
 
 const OPEN_MODAL_TEXT = 'Open Modal';
 const USERNAME = 'test_username1';
