+++ webui/react/src/pages/Cluster/ClusterOverallStats.test.tsx
@@ -13,15 +13,11 @@ vi.mock('services/api', () => ({
   getResourcePools: () => Promise.resolve({}),
 }));
 
-jest.mock('index', () => {
-  return {
-    router: {
-      navigate: (path: string) => {
-        global.window.history.pushState({}, '', path);
-      },
-    },
-  };
-});
+jest.mock('router', () => ({
+  navigate: (path: string) => {
+    global.window.history.pushState({}, '', path);
+  },
+}));
 
 const setup = () => {
   const view = render(
