+++ webui/react/src/pages/InteractiveTask.test.tsx
@@ -22,15 +22,11 @@ vi.mock('react-router-dom', () => ({
   useRouteMatch: () => ({ url: '/company/company-id1/team/team-id1' }),
 }));
 
-jest.mock('index', () => {
-  return {
-    router: {
-      navigate: (path: string) => {
-        global.window.history.pushState({}, '', path);
-      },
-    },
-  };
-});
+jest.mock('router', () => ({
+  navigate: (path: string) => {
+    global.window.history.pushState({}, '', path);
+  },
+}));
 
 const InteractiveTaskPageContainer: React.FC = () => {
   useEffect(() => {
