+++ webui/react/src/services/apiConfig.test.ts
@@ -1,14 +1,10 @@
 import * as utils from './apiConfig';
 
-jest.mock('index', () => {
-  return {
-    router: {
-      navigate: (path: string) => {
-        global.window.history.pushState({}, '', path);
-      },
-    },
-  };
-});
+jest.mock('router', () => ({
+  navigate: (path: string) => {
+    global.window.history.pushState({}, '', path);
+  },
+}));
 
 describe('apiConfig', () => {
   describe('getUserIds', () => {
