+++ webui/react/src/shared/utils/routes.test.ts
@@ -2,20 +2,16 @@ import React from 'react';
 
 import { reactHostAddress } from 'routes/utils';
 
-import { router } from '../../index';
+import router from '../../router';
 
 import type { AnyMouseEvent } from './routes';
 import * as routes from './routes';
 
-jest.mock('index', () => {
-  return {
-    router: {
-      navigate: (path: string) => {
-        global.window.history.pushState({}, '', path);
-      },
-    },
-  };
-});
+jest.mock('router', () => ({
+  navigate: (path: string) => {
+    global.window.history.pushState({}, '', path);
+  },
+}));
 
 const setup = (url = '/', base = 'http://www.example.com') => {
   const newUrl = new URL(url, base);
