+++ webui/react/src/shared/utils/routes.ts
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { router } from '../../index';
+import router from '../../router';
 
 import { clone } from './data';
 import rootLogger from './Logger';
