+++ webui/react/src/utils/browser.test.ts
@@ -1,14 +1,10 @@
 import * as utils from './browser';
 
-jest.mock('index', () => {
-  return {
-    router: {
-      navigate: (path: string) => {
-        global.window.history.pushState({}, '', path);
-      },
-    },
-  };
-});
+jest.mock('router', () => ({
+  navigate: (path: string) => {
+    global.window.history.pushState({}, '', path);
+  },
+}));
 
 describe('Browser Utilities', () => {
   describe('cookies', () => {
