+++ webui/react/src/utils/job.test.ts
@@ -3,15 +3,11 @@ import { CommandType, Job, JobType } from 'types';
 
 import * as utils from './job';
 
-jest.mock('index', () => {
-  return {
-    router: {
-      navigate: (path: string) => {
-        global.window.history.pushState({}, '', path);
-      },
-    },
-  };
-});
+jest.mock('router', () => ({
+  navigate: (path: string) => {
+    global.window.history.pushState({}, '', path);
+  },
+}));
 
 describe('Job Utilities', () => {
   describe('jobTypeIconName', () => {
